backend/requirements.txt
fastapi
uvicorn
librosa
numpy
python-dotenv
openai>=1.0.0
elevenlabs==0.2.24
pydantic==1.10.12
